refactor(tests): split cli fakes
